Struct rusoto_mgh::MigrationHubClient

pub struct MigrationHubClient { /* fields omitted */ }

A client for the AWS Migration Hub API.

Methods

impl MigrationHubClient

pub fn new(region: Region) -> MigrationHubClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> MigrationHubClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,