Crate rusoto_mgh

The AWS Migration Hub API methods help to obtain server and application migration status and integrate your resource-specific migration tool by providing a programmatic interface to Migration Hub.

If you're using the service, you're probably looking for MigrationHubClient and MigrationHub.

Structs

AssociateCreatedArtifactRequest

AssociateCreatedArtifactResult

AssociateDiscoveredResourceRequest

AssociateDiscoveredResourceResult

CreateProgressUpdateStreamRequest

CreateProgressUpdateStreamResult

CreatedArtifact
An ARN of the AWS cloud resource target receiving the migration (e.g., AMI, EC2 instance, RDS instance, etc.).

DeleteProgressUpdateStreamRequest

DeleteProgressUpdateStreamResult

DescribeApplicationStateRequest

DescribeApplicationStateResult

DescribeMigrationTaskRequest

DescribeMigrationTaskResult

DisassociateCreatedArtifactRequest

DisassociateCreatedArtifactResult

DisassociateDiscoveredResourceRequest

DisassociateDiscoveredResourceResult

DiscoveredResource
Object representing the on-premises resource being migrated.

ImportMigrationTaskRequest

ImportMigrationTaskResult

ListCreatedArtifactsRequest

ListCreatedArtifactsResult

ListDiscoveredResourcesRequest

ListDiscoveredResourcesResult

ListMigrationTasksRequest

ListMigrationTasksResult

ListProgressUpdateStreamsRequest

ListProgressUpdateStreamsResult

MigrationHubClient
A client for the AWS Migration Hub API.

MigrationTask
Represents a migration task in a migration tool.

MigrationTaskSummary
MigrationTaskSummary includes MigrationTaskName, ProgressPercent, ProgressUpdateStream, Status, and UpdateDateTime for each task.

NotifyApplicationStateRequest

NotifyApplicationStateResult

NotifyMigrationTaskStateRequest

NotifyMigrationTaskStateResult

ProgressUpdateStreamSummary
Summary of the AWS resource used for access control that is implicitly linked to your AWS account.

PutResourceAttributesRequest

PutResourceAttributesResult

ResourceAttribute

Attribute associated with a resource.

Note the corresponding format required per type listed below:

IPV4

x.x.x.x

where x is an integer in the range [0,255]

IPV6

y : y : y : y : y : y : y : y

where y is a hexadecimal between 0 and FFFF. [0, FFFF]

MAC_ADDRESS

^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$

FQDN

^[^<>{}\\/?,=\p{Cntrl}]{1,256}$

Task
Task object encapsulating task information.

Enums

AssociateCreatedArtifactError
Errors returned by AssociateCreatedArtifact

AssociateDiscoveredResourceError
Errors returned by AssociateDiscoveredResource

CreateProgressUpdateStreamError
Errors returned by CreateProgressUpdateStream

DeleteProgressUpdateStreamError
Errors returned by DeleteProgressUpdateStream

DescribeApplicationStateError
Errors returned by DescribeApplicationState

DescribeMigrationTaskError
Errors returned by DescribeMigrationTask

DisassociateCreatedArtifactError
Errors returned by DisassociateCreatedArtifact

DisassociateDiscoveredResourceError
Errors returned by DisassociateDiscoveredResource

ImportMigrationTaskError
Errors returned by ImportMigrationTask

ListCreatedArtifactsError
Errors returned by ListCreatedArtifacts

ListDiscoveredResourcesError
Errors returned by ListDiscoveredResources

ListMigrationTasksError
Errors returned by ListMigrationTasks

ListProgressUpdateStreamsError
Errors returned by ListProgressUpdateStreams

NotifyApplicationStateError
Errors returned by NotifyApplicationState

NotifyMigrationTaskStateError
Errors returned by NotifyMigrationTaskState

PutResourceAttributesError
Errors returned by PutResourceAttributes

Traits

MigrationHub
Trait representing the capabilities of the AWS Migration Hub API. AWS Migration Hub clients implement this trait.